rename appDeploy and refactor AC controller
