+++ docs/examples/cloneset-rollout/app-rollout-finish.yaml
