+++ docs/examples/cloneset-rollout/app-rollout-pause.yaml
