+++ docs/examples/cloneset-rollout/app-rollout-revert.yaml
