+++ docs/examples/cloneset-rollout/app-source-prep.yaml
