+++ docs/examples/cloneset-rollout/app-source.yaml
