+++ docs/examples/cloneset-rollout/app-target-done.yaml
