+++ docs/examples/cloneset-rollout/app-target.yaml
