+++ docs/examples/cloneset-rollout/clonesetDefinition.yaml
