+++ pkg/controller/core.oam.dev/v1alpha2/applicationrollout/applicationrollout_controller.go
@@ -1,4 +1,4 @@
-package applicationdeployment
+package applicationrollout
 
 import (
 	"context"
