+++ pkg/controller/core.oam.dev/v1alpha2/applicationrollout/suite_test.go
@@ -1,4 +1,4 @@
-package applicationdeployment
+package applicationrollout
 
 import (
 	"os"
